refactor(frontend): extract YieldFilterBar into its own module
Move the self-contained YieldFilterBar sub-component out of the large
YieldOverview page file into src/pages/YieldFilterBar.tsx, shrinking the
page by ~130 lines. The shared ALL_CULTURES constant and YieldCultureMeta
type move into yieldOverviewUtils so both the page and the filter bar import
them without a circular dependency. No behavior change.
